Add fish function for converting PDFs to JPGs.
If I need to scan a document, I use Google Drive's scan feature on my
phone which creates a PDF. But sometimes you need an image.
This converts the scanned PDF to a sharp and readable JPG.

Although the size of the JPG isn't that large (relative to the PDF),
it may be worth adding a compression option to reduce the size.
Something to consider...