Use a better find command and fix it for OS X
The BSD version of find that ships with OS X requires a search path.
In this case, I had to add a '.' so it searches the current directory,
because that's not default behaviour if the path isn't specified.

I also switched the find command to the one I had commented previously.
The reason is that this version also finds your .gitconfig which you'll
probably want to include as part of your dotfiles.