Optimization: basic test to see if in a git repo
The time saved by preventing a call to `git` is significant.

The extra time for the call to `test` is insignificant on top of the
time it takes to call `git` (when a repo does exist).

Note that this is a basic test. It can be fooled when an empty
.git directory exists (that is, not a real git repo). In those cases,
the redirection of stderr will handle the failure. There are ways
to avoid the redirection (e.g. enhance the test with a fallback to git
rev-parse) but then `git` will be called twice in true positive cases
and that's not ideal. I prefer the single call with redirection.

1. IS NOT a git repo, redirecting stderr
↪ bash -c 'time for i in {1..2000}; do git symbolic-ref --short HEAD &> /dev/null; done'

real	0m6.624s
user	0m1.851s
sys	0m2.965s

2. IS NOT a git repo, basic short-circuit test for git dir
↪ bash -c 'time for i in {1..2000}; do test -d .git && git symbolic-ref --short HEAD &> /dev/null; done'

real	0m0.019s
user	0m0.016s
sys	0m0.002s

3. IS a git repo, basic short-circuit test for git dir
↪ bash -c 'time for i in {1..2000}; do test -d .git && git symbolic-ref --short HEAD &> /dev/null; done'

real	0m8.545s
user	0m2.580s
sys	0m4.096s

4. IS a git repo, redirecting stderr
↪ bash -c 'time for i in {1..2000}; do git symbolic-ref --short HEAD &> /dev/null; done'

real	0m8.248s
user	0m2.551s
sys	0m3.845s