Move git-* functions to bin/
The bin/ directory is meant to be in your PATH, and whenever there is a
git-some-name script in your PATH, git will automatically pick it up.

I tried moving these functions to .config/fish/functions/git/ but that
wasn't idead because the autoloading of fish functions didn't seem to
work for those git functions anymore.