Fix rbenv check
Went with testing for the existence of a file initially because it's
more efficient - should keep config.fish as lightweight as possible
since it's executed with every new shell instance.

However, depending on how you install rbenv, that file may not exist.
For example, it does exist when you follow clone instructions, but not
when using homebrew.

So the check could be false negative and abort initializing rbenv.