Add fish helpers to quickly run WIP Elixir tests
The 'mtest' function allows you to select a modified test file in the
Elixir umbrella (with cwd at the umbrella level or lower) to run.
It also shows the diff of the current file in the preview pane.

More convenient than me having to do the following during TDD:
1. git status
2. copy path for a test file
3. remove the 'apps/<app>' prefix so I can run it from the root level
4. paste path and prefix 'mix test --only x' or some variation of it

(I prefer to run my tests in an external terminal rather than in my
editor. Used to do it in the editor for a long time, but it's just not
as good of an experience.)

Still need to replace the hardcoded app names with a token regex.