Make wip-tests function more robust
Using it with short status meant there were more cases to handle,
because a file can be in three different modified stages. E.g.

 M   file1
 MM  file2
  M  file3

By using the normal status, it's easier to match on the files no matter
what modified state they're in. But we have to remove duplicates as a
file can appear multiple times due to being in multiple modified stages.