Allow mtest to be used with a test path passed as an argument
Sometimes I want to run a test module without having made changes to it
yet. In that case I can now call the function as "mtest apps/<testpath>'
and not select anything during the prompt, and it will run the given
test. This is possible thanks to including the $opts argument in the
normalisation process, since it removes the umbrella app prefix and any
leading whitespace that would interfere with 'mix test' understanding.