:fish: Add remove_from_path function
Had a lot of paths set directly on PATH (probably from 3rd party
install scripts) rather than via fish_user_path. Included duplicates.

Note we don't need to set the scope or export, because if a variable
already exists, it uses the same scope and export rules on update.

But I was curious. According to `set --show PATH`, it's set as global
scope and exported.