Output file even if same or larger size
I haven't been able to produce a larger file size yet but producing the
same file size is trivial by giving it an already compressed image.

We still want to run the command in that case to avoid any confusion for
the end user wondering why the file's timestamp hasn't been updated or
why a new file wasn't created when using the --output option.