Fix function detection for compression funcs
When fish encounters a command, it attempts to autoload a function
matching the command by looking for a file with the matching name in
`~/.config/fish/functions/`. That's why the file-specific compressions
functions like `compress_pdf` weren't loading until `compress_all` was.

We fix that by extracting the functions to their own files. Hence we
also renamed the functions to all have the same `compress_` prefix so
that they're grouped alphabetically.