Add userscript for headless gmail print
Started by creating a Safari Web Extension but that was a mistake.
Thought it would be trivial as I've written a handful of Chrome
extensions, but the Safari and Xcode ecosystem is full of dragons.

Instead decided to use https://github.com/quoid/userscripts. an open
source extension that lets you load GM-compatible user scripts similar
to Greasemonkey et al.