Remove deprecated path functions
We can use fish_add_path and fish_remove_path from now on.

Note that fish_remove_path doesn't give you the granularity to choose
if you want to remove from $PATH or $fish_user_paths but I don't think
there will ever be a scenario where we want to remove from one and not
the other. And if there is, can add an option to that function.